Updated version information and cleaned up. Fixed Margins. Move common code.
