+++ Economie/App.xaml
@@ -2,9 +2,5 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              StartupUri="MainWindow.xaml">
-    <Application.Resources>
-        <ResourceDictionary>
-            <Style TargetType="{x:Type Rectangle}" />
-        </ResourceDictionary>
-    </Application.Resources>
+    
 </Application>
