+++ Economie/BeamerWindows/BeamerWindow.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Windows;
+using System.Windows.Input;
 
 namespace Nobel.Economie.BeamerWindows {
     public abstract class BeamerWindow : Window, IBeamerWindow {
@@ -15,7 +16,25 @@ namespace Nobel.Economie.BeamerWindows {
                     allowClosing = value;
             }
         }
+        #region Event Handlers
+        private void Window_Closing(object sender, System.ComponentModel.CancelEventArgs e) {
+            e.Cancel = !this.AllowClosing;
+        }
 
+        private void Window_KeyDown(object sender, KeyEventArgs e) {
+            if (e.Key == Key.F11) {
+                if (WindowState != System.Windows.WindowState.Maximized) {
+                    WindowStyle = System.Windows.WindowStyle.None;
+                    WindowState = System.Windows.WindowState.Maximized;
+                    ResizeMode = System.Windows.ResizeMode.NoResize;
+                } else {
+                    WindowStyle = System.Windows.WindowStyle.SingleBorderWindow;
+                    WindowState = System.Windows.WindowState.Normal;
+                    ResizeMode = System.Windows.ResizeMode.CanResize;
+                }
+            }
+        }
+        #endregion
 
         public abstract void updateDisplay(ref Dictionary<string, long> points);
         public abstract void updateTimeslots(ref List<Timeslot> timeslots, ref Dictionary<string, int> pointsources, DateTime ecoStart, TimeSpan timeslotLen);
