+++ Economie/BeamerWindows/ScrollerBeamerWindow.xaml.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Animation;
 
@@ -124,24 +123,7 @@ namespace Nobel.Economie.BeamerWindows {
 		{
             checkAnimation();		
 		}
-        private void Window_KeyDown(object sender, KeyEventArgs e) {
-            if (e.Key == Key.F11) {
-                if (WindowState != System.Windows.WindowState.Maximized) {
-                    WindowStyle = System.Windows.WindowStyle.None;
-                    WindowState = System.Windows.WindowState.Maximized;
-                    ResizeMode = System.Windows.ResizeMode.NoResize;
-                } else {
-                    WindowStyle = System.Windows.WindowStyle.SingleBorderWindow;
-                    WindowState = System.Windows.WindowState.Normal;
-                    ResizeMode = System.Windows.ResizeMode.CanResize;
-                }
-                checkAnimation();
-            }
 
-        }
-        private void Window_Closing(object sender, System.ComponentModel.CancelEventArgs e) {
-            e.Cancel = !this.AllowClosing;
-        }
         private void Window_SizeChanged(object sender, SizeChangedEventArgs e)
 		{
             checkAnimation();
